Ok, fix the clock. Add mux barrier. Thread in BaB...
These are things we have kept back for as long as possible but we can hold it no longer. Notably, the thread isn't a full solution to the blocking issues but it has come to our attention that the Calvin philosophy requires us to thread out (although not this way). If parallellized on a different core though, it is not necessarily good for our case...